@@ -4,12 +4,12 @@ version: 2.0.1
 homepage: https://github.com/skkallayath/photofilters
 
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: ">=3.1.3 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  image: ^3.0.2
+  image: 4.1.3
   meta: ^1.1.6
   path_provider: ^2.0.2
 
@@ -17,7 +17,7 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
   test:
-  image_picker: ^0.6.0+10
+  image_picker: ^1.0.4
   path: ^1.6.2
 
 # For information on the generic Dart part of this file, see the
