@@ -12,7 +12,7 @@ abstract class ColorSubFilter extends SubFilter {
 ///The [ColorFilter] class to define a Filter which will applied to each color, consists of multiple [SubFilter]s
 class ColorFilter extends Filter {
   List<ColorSubFilter> subFilters;
-  ColorFilter({String name, String nameHindi})
+  ColorFilter({String? name, String? nameHindi})
       : subFilters = [],
         super(name: name, nameHindi: nameHindi);
 
